requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
python-dotenv==0.19.0
email-validator==1.1.3
Pillow==8.3.2